% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_areanames.R
\name{available_areanames}
\alias{available_areanames}
\title{List of areas that are present in a plotted brain slice}
\usage{
available_areanames(composite, slice = NULL)
}
\arguments{
\item{composite}{Object from either SpatialEnrichment or CreateBrain}

\item{slice}{Section of the brain to be queried}
}
\value{
all areas present in the brain section of interest
}
\description{
provides all brain areas within a slice for a refset.
}
\examples{
#brainImageR:::loadworkspace()
##Load in a gene set
data(vth)
##calculate spatial enrichment
#composite <- SpatialEnrichment(vth, reps = 20, refset = "developing")
#available_areanames(composite, slice = 5)

}
