% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boot.R
\name{Boot}
\alias{Boot}
\title{Calculate significance of gene set enrichment}
\usage{
Boot(composite)
}
\arguments{
\item{composite}{Object returned from SpatialEnrichment.}
}
\value{
spatiotemporal prediction
}
\description{
\code{Boot} bootstraps the likelihood of the observed enrichment
}
\examples{
#brainImageR:::loadworkspace()
##First put together a gene list, or load in the default vth dataset
data(vth)
##Calculate the spatial enrichment.
#composite <- SpatialEnrichment(vth, reps = 20, refset = "developing")
#tissueExp1 <- composite@tissueExp1
#random.matrix <- composite@random.matrix
##Calculate the significance estimates
#boot <- Boot(composite)
#boot <- boot[order(boot$FC, decreasing=TRUE),]
#head(boot)

}
