% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainMap.R
\name{BrainMap}
\alias{BrainMap}
\title{Internal- Overlaps regional enrichment into a single section}
\usage{
BrainMap(dim, tissueExp, Abrev, Files, slice, refset = c("developing",
    "adult"))
}
\arguments{
\item{dim}{numeric dimensions of the original image}

\item{tissueExp}{counts of genes per tissue, from SpatialEnrichment.}

\item{Abrev}{character of all regions in the given section}

\item{Files}{character of tiff images for each region}

\item{slice}{integer of current slice}

\item{refset}{character of reference brain map}
}
\value{
returns a matrix weighted by the gene overlap
}
\description{
\code{BrainMap} Merges maps from reColor
}
\examples{
##Internal to brainImageR,called within CreateBrain
#brainImageR:::loadworkspace()
##First load in a gene set
data(vth)
##calculate the spatial enrichment
#composite <- SpatialEnrichment(vth, reps = 20, refset = "developing")
#tissueExp <- Boot(composite)

##Select the slice of interest
#slice <- 4
#Files <- .cache[["EH1434"]][[slice]]
#dim <- .cache[["EH1436"]][[slice]]

#Select the region of interest
#Abrev <- .cache[["EH1438"]]
#abrev <- "VZ"


# map <- BrainMap(dim = dim ,
# tissueExp = tissueExp,
#  Abrev = Abrev, Files = Files,
#  slice = slice,
# refset = "developing")


}
