% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainrange.R
\name{brainrange}
\alias{brainrange}
\title{brainrange}
\usage{
brainrange(first = 0, last = 1, by = 1)
}
\arguments{
\item{first}{starting value number}

\item{last}{ending value}

\item{by}{amount to move by in sequence}
}
\value{
vector of numbers
}
\description{
creates a sequence of numbers from first to last by the given interval
}
\examples{
brainrange(1,10,0.5)

}
