% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comp.R
\docType{class}
\name{Comp-class}
\alias{Comp-class}
\alias{Comp}
\title{Set class Comp}
\description{
Set class Comp
}
\examples{
comp <- methods::new(Class="Comp",
tissueExp1 = c(10,12),
tissueExp2 =  table(c(rep("A",5), rep("B",5))),
composite = matrix(0,nrow=10,ncol=10),
random.matrix = data.frame(matrix(0,nrow=10,ncol=10)),
refset = "developing"
)
}
