% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBrain.R
\name{CreateBrain}
\alias{CreateBrain}
\title{Overlap spatial enrichment information and anatomical organization}
\usage{
CreateBrain(composite, boot, slice, pcut = 0.05)
}
\arguments{
\item{composite}{Comp object returned from SpatialEnrichment}

\item{boot}{result from testEnrich including significance estimates}

\item{slice}{integer brain section}

\item{pcut}{numeric padj filter.}
}
\value{
Comp object
}
\description{
\code{CreateBrain} convert spatial enrichment into anatomical coordinates.
}
\examples{
#brainImageR:::loadworkspace()
##First put together a gene list, or load in the default vth dataset
data(vth)
##Calculate the spatial enrichment.
#composite <- SpatialEnrichment(vth, reps = 20, refset = "developing")
#tissueExp1 <- composite@tissueExp1
#random.matrix <- composite@random.matrix
##Calculate the significance estimates
#boot <- testEnrich(composite)
##Color the brain section of interest with enrichment
#composite <- CreateBrain(composite, boot, slice = 6, pcut = 0.05)
##Plot the brain
#PlotBrain(composite, Breaks = 12)

}
