% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{dat}
\alias{dat}
\title{in vitro temporal data of human neural progenitor cells and neurons}
\format{10193 X 6 data.frame}
\usage{
dat
}
\description{
in vitro temporal data of human neural progenitor cells and neurons
}
\keyword{datasets}
