% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneLists.R
\docType{data}
\name{GeneLists}
\alias{GeneLists}
\title{Lists of genes associated with each brain region}
\format{Names of each list object is the corresponding brain region.}
\usage{
GeneLists
}
\description{
Lists of genes associated with each brain region
}
\keyword{datasets}
