% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGenes.R
\name{GetGenes}
\alias{GetGenes}
\title{GenGenes}
\usage{
GetGenes(genes, composite, tissue_abrev = NULL)
}
\arguments{
\item{genes}{Query gene list.}

\item{composite}{Result from SpatialEnrichment}

\item{tissue_abrev}{The tissue of interest.}
}
\value{
Gene overlap between query and tissue of interest
}
\description{
\code{GetGenes} returns the genes that are expressed within a given tissue
}
\examples{
#brainImageR:::loadworkspace()
##First put together a gene list, or load in the default vth dataset
data(vth)
##Calculate the spatial enrichment.
#composite <- SpatialEnrichment(vth, reps = 20, refset = "developing")
##Ask which genes are present in any given tissue.
#available_areanames(composite, slice = 4)
#vth_in_VZ <- GetGenes(vth, composite, tissue_abrev = "VZ")

}
