% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InABA.R
\name{InABA}
\alias{InABA}
\title{Quick search for presence of genes in the ABA list}
\usage{
InABA(genes, refset = "developing")
}
\arguments{
\item{genes}{genes to search}

\item{refset}{reference brain map. developing (default) or adult}
}
\value{
returns the list of genes that are also present in the ABA dataset
}
\description{
\code{InABA} checks for the gene name in the ABA dataset
}
\examples{

#brainImageR:::loadworkspace()
##First load in a gene set
data(vth)
##Then query the dataset to see which genes are present
#vth_in <- InABA(vth)
#head(vth_in)
#length(vth_in) / length(vth)

}
