% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBrain.R
\name{PlotBrain}
\alias{PlotBrain}
\title{Color and Plot the SGSE image}
\usage{
PlotBrain(composite, legend = TRUE)
}
\arguments{
\item{composite}{Object returned from CreateBrain.}

\item{legend}{whether the legend should be included. Default = TRUE}
}
\value{
plots the SGSE brain image
}
\description{
\code{PlotBrain} Plots CreateBrain.
}
\details{
PlotBrain plots your spatial gene set enrichment image.
}
\examples{

##First put together a gene list, or load in the default vth dataset
#brainImageR:::loadworkspace()
data(vth)
##Calculate the spatial enrichment.
#composite <- SpatialEnrichment(vth, reps = 20, refset = "developing")
##Calculate the significance of the gene set enrichment
#boot <- Boot(composite)
##Color the brain section of interest
#composite <- CreateBrain(composite, boot, slice = 5, pcut = 0.05)
##Plot the brain
#PlotBrain(composite)
}
