% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotEnrich.R
\name{PlotEnrich}
\alias{PlotEnrich}
\title{PlotEnrich}
\usage{
PlotEnrich(boot)
}
\arguments{
\item{boot}{Comp object returned from the testEnrich function}
}
\value{
Spatial enrichment plot
}
\description{
A quick plot to assess the enrichments returned from testEnrich.
Gene overlap calculated by random chance is plotted on the
x-axis and the gene overlap from the query set on the y-axis. Each
dot represents an individual microdissected tissue. Note that
the signficance estimate is only dependent on the randomly generated
overlaps if the p-values were calculated with the bootstrap procedure.
}
\examples{
#brainImageR:::loadworkspace()
##First put together a gene list, or load in the default vth dataset
data(vth)
##Calculate the spatial enrichment
#composite <- SpatialEnrichment(vth, 20, "developing")
#tissueExp1 <- composite@tissueExp1
#random.matrix <- composite@random.matrix
##Calculate the significance estimates
#boot <- testEnrich(composite)
#PlotEnrich(boot)

}
