% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPred.R
\name{PlotPred}
\alias{PlotPred}
\title{Plot your temporal predictions}
\usage{
PlotPred(time)
}
\arguments{
\item{time}{Object returned from predict_time}
}
\value{
prediction plot
}
\description{
\code{PlotPred} Plots the temporal predictions.
}
\examples{
#brainImageR:::loadworkspace()
##Load in data
data(dat)
##predict time
#time <- predict_time(dat)
##plot the predictions
#PlotPred(time)
}
