% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pred.R
\docType{class}
\name{Pred-class}
\alias{Pred-class}
\alias{Pred}
\title{Pred}
\description{
keeps track of parameters and results from predict_time
}
\section{Slots}{

\describe{
\item{\code{pred_age}}{data.frame of results from predict_time}

\item{\code{model}}{randomForest model}

\item{\code{minage}}{minimum age filter from predict_time}

\item{\code{maxage}}{maximum age filter from predict_time}

\item{\code{tissue}}{tissue filter from predict_time}
}}

\examples{
prep <- methods::new(Class="Pred",
pred_age = data.frame(matrix(0,nrow=10,ncol=10)),
model =  list(c(rep("A",5), rep("B",5))),
minage = 8,
maxage = 2120,
tissue = "HIP"
)
}
