% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.time.R
\name{predict_time}
\alias{predict_time}
\title{Predict developmental time from gene expression data}
\usage{
predict_time(dat = NULL, genelist = NULL, minage = 8,
  maxage = 2120, tissue = NULL, minrsq = 0.6)
}
\arguments{
\item{dat}{Normalized expression matrix}

\item{genelist}{Optional: restrict analysis to gene list}

\item{minage}{min pcw of the reference set. default = 8}

\item{maxage}{max pcw of the reference set. default = 2120}

\item{tissue}{Optional: restrict analysis to tissue (available)}

\item{minrsq}{(range 0-1) model leniency. default = 0.5.}
}
\value{
spatiotemporal predictions.
}
\description{
Predict human developmental time from expression dataset
}
\examples{
#brainImageR:::loadworkspace()
##Load in the data
data(dat)
##predict time
#time <- predict_time(dat)
}
