% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomTissueSummary.R
\name{RandomTissueSummary}
\alias{RandomTissueSummary}
\title{Generate a random overlap}
\usage{
RandomTissueSummary(i, genes, samplesize, refset = c("developing",
  "adult"))
}
\arguments{
\item{i}{current iteration}

\item{genes}{all genes to sample from}

\item{samplesize}{sample size to select from gene list}

\item{refset}{reference map. developing (default) or adult}
}
\value{
returns a list of the random gene overlap for each tissue
}
\description{
random overlap for background correction and bootstrapping
}
\examples{
##Internal to brainImageR, called within SpatialEnrichment
#brainImageR:::loadworkspace()
##First load in a gene set
data(vth)
#tissueExp <- RandomTissueSummary(1, vth, 20)


}
