% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{range}
\alias{range}
\title{range}
\usage{
range(first = 0, last = 1, by = 1)
}
\arguments{
\item{first}{starting value number}

\item{last}{ending value}

\item{by}{amount to move by in sequence}
}
\value{
vector of numbers
}
\description{
replaces seq to pass bioccheck
}
\examples{
range(1,10,0.5)
}
