% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reColor.R
\name{reColor}
\alias{reColor}
\title{Color in the brain images based on enrichment values}
\usage{
reColor(i, slice, tissueExp, dim, Abrev, Files, refset = c("developing",
    "adult"))
}
\arguments{
\item{i}{tissue region from within the specified rostral-caudal section}

\item{slice}{current slice}

\item{tissueExp}{tissueExp1 from SpatialEnrichment.}

\item{dim}{Original dimensions of the image}

\item{Abrev}{list of tissue regions}

\item{Files}{list of tiff images}

\item{refset}{reference map. developing (default) or adult}
}
\value{
returns genes counts for each tissue
}
\description{
\code{reColor} quantifies the presence of a gene list within each tissue
}
\examples{

#brainImageR:::loadworkspace()
##First load a gene set
data(vth)
##calculate spatial enrichment
#composite <- SpatialEnrichment(vth,20,"developing")
#boot <- Boot(composite)
#subboot <- c(boot[boot$pvalue < 0.05 & is.finite(boot$FC), "FC"])
#names(subboot) <- rownames(boot[boot$pvalue < 0.05 & is.finite(boot$FC), ])
#tissueExp <- subboot
##Select the slice of interest
#slice <- 4
#Files <- .cache[["EH1434"]][[slice]]
#dim <- .cache[["EH1436"]][[slice]]
##Select the region of interest
#Abrev <- .cache[["EH1438"]][[4]]
#abrev <- "VZ"
#tmp <- reColor(abrev, slice, tissueExp, dim, Abrev, Files)


}
