% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialEnrichment.R
\name{SpatialEnrichment}
\alias{SpatialEnrichment}
\title{Calculate the presence of your gene set within each brain region}
\usage{
SpatialEnrichment(genes, background = NULL, reps = 10,
    refset = "developing")
}
\arguments{
\item{genes}{query gene set}

\item{background}{background gene list, default = NULL (uses all ABA genes)}

\item{reps}{replicates for bootstrap, default = 10}

\item{refset}{reference brain map. developing (default) or adult}
}
\value{
"Comp" object
}
\description{
Calculates the presence of gene set within each region
}
\examples{
#brainImageR:::loadworkspace()
##First load in a gene set
data(vth)
##Then calculate the spatial enrichment
#composite <- SpatialEnrichment(vth,20,"developing")
}
