% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testEnrich.R
\name{testEnrich}
\alias{testEnrich}
\title{Calculate significance of gene set enrichment}
\usage{
testEnrich(composite, method = "fisher")
}
\arguments{
\item{composite}{Object returned from SpatialEnrichment.}

\item{method}{either "bootstrap" or "fisher"}
}
\value{
spatiotemporal prediction
}
\description{
\code{testEnrich} test the enrichment of the observed enrichment
}
\examples{
#brainImageR:::loadworkspace()
##First put together a gene list, or load in the default vth dataset
data(vth)
##Calculate the spatial enrichment.
#composite <- SpatialEnrichment(vth, reps = 20, refset = "developing")
##Calculate the significance estimates
#boot <- testEnrich(composite)
#boot <- boot[order(boot$FC, decreasing=TRUE),]
#head(boot)

}
