% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tis_in_region.R
\name{tis_in_region}
\alias{tis_in_region}
\title{List of regions that are supported by given tissue}
\usage{
tis_in_region(composite, tissue_abrev)
}
\arguments{
\item{composite}{object returned from SpatialEnrichment or CreateBrain}

\item{tissue_abrev}{abreviation of the microdissected tissue of interest.}
}
\value{
general brain areas
}
\description{
Brain areas supported by tissue of interest. Opposite=tis_set()
}
\examples{
#brainImageR:::loadworkspace()
##Load in a gene set
data(vth)
##calculate spatial enrichment
#composite <- SpatialEnrichment(vth, reps = 20, refset = "developing")
#tis_in_region(composite, "LHAa")
}
