% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tis_set.R
\name{tis_set}
\alias{tis_set}
\title{List of tissues that support a given region in the brain plot}
\usage{
tis_set(composite, area.name, slice)
}
\arguments{
\item{composite}{object returned from SpatialEnrichment or CreateBrain}

\item{area.name}{abreviation of the brain area of interest.}

\item{slice}{section of the brain to query (1-10)}
}
\value{
all areas supported by the tissue
}
\description{
Tissues that support brain area. Opposite=tis_in_region()`
}
\examples{
#brainImageR:::loadworkspace()
##Load in a gene set
data(vth)
##calculate spatial enrichment
#composite <- SpatialEnrichment(vth, reps = 20, refset = "developing")
#get the set of tissues that are present within a given region
#tis_set(composite, area.name = "Pu", slice = 6)
}
