% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissuemean.R
\name{tissuemean}
\alias{tissuemean}
\title{Internal-Take mean of all samples per tissue}
\usage{
tissuemean(region, LIST)
}
\arguments{
\item{region}{region of interest}

\item{LIST}{full set of all values}
}
\value{
returns the mean of region for the whole list
}
\description{
mean of all samples per tissue
}
\examples{
x <- 1

}
