% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TissueSummary.R
\name{TissueSummary}
\alias{TissueSummary}
\title{Internal-Identify the number of genes expressed in each tissue}
\usage{
TissueSummary(genes, refset = c("developing", "adult"))
}
\arguments{
\item{genes}{Genes to query for tissue location}

\item{refset}{reference map. developing (default) or adult}
}
\value{
returns genes counts for each tissue
}
\description{
.
Quantifies the presence of gene list within each tissue
}
\examples{

#Internal to brainImageR, used within SpatialEnrichment
#First load in a gene set
#brainImageR:::loadworkspace()
data(vth)
#tissueExp <- TissueSummary(vth, refset = "developing")

}
