% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vth.R
\docType{data}
\name{vth}
\alias{vth}
\title{VTH genes identified from Allen Brain Atlas developing human}
\format{A vector with 1389 gene names}
\usage{
vth
}
\description{
Ventral thalamus genes were identified using the differential
search tool on the Allen Brain Atlas brain span (prenatal brain).
The ventral thalamus was contrasted to neighboring regions to identify a
gene list that was enriched within the vth.
}
\keyword{datasets}
