% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comp.R
\docType{class}
\name{Comp-class}
\alias{Comp-class}
\alias{Comp}
\title{Comp keeps track of Spatial Enrichment calculations and parameters}
\description{
Comp tracks the parameters and calculations throughout spatial
gene set enrichment.
}
\section{Slots}{

\describe{
\item{\code{genes}}{character vector of query genes}

\item{\code{tissueExp1}}{named numeric vector of query gene count in tissues}

\item{\code{tissueExp2}}{named numeric vector of avg.}

\item{\code{random.matrix}}{matrix of overlap at random, size = boot replicates}

\item{\code{refset}}{character noting developing or adult reference

 background gene count in tissues}

\item{\code{composite}}{composite image matrix}
}}

\examples{
comp <- methods::new(Class="Comp",
genes = c("a","b"),
tissueExp1 = c(10,12),
tissueExp2 =  c(10,13),
composite = matrix(0,nrow=10,ncol=10),
random.matrix = data.frame(matrix(0,nrow=10,ncol=10)),
refset = "developing"
)
}
