% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{dat}
\alias{dat}
\title{In vitro temporal data of human neural progenitor cells and neurons}
\format{10193 X 6 data.frame}
\usage{
dat
}
\description{
RNA-seq data: Induced pluripotent stem cells were differentiated
into neural progenitor cells (NPC.1, NPC.2, NPC.3, NPC.4) that were patterned
for the forebrain lineage. Neurons were differentiated from
these neural progenitor cells (Neurons.5, Neurons.6).
RNA was collected from both NPCs and neurons,
poly(A) cDNA libraries were generated, and sequenced
on an Illumina HiSeq 2500.
RNA-seq data was aligned to the human Hg19 reference, and counts were
normalized into log2(TPM + 1) values. Gene names are in human symbol format.
}
\keyword{datasets}
