% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hipp.R
\docType{data}
\name{hipp}
\alias{hipp}
\title{Hippocampal genes using the adult human brain as reference}
\format{a vector with 1302 gene names}
\usage{
hipp
}
\description{
Hippocampal genes using the adult human brain as reference
}
\keyword{datasets}
