% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBrain.R
\name{PlotBrain}
\alias{PlotBrain}
\title{Color and Plot the SGSE image}
\usage{
PlotBrain(composite, legend = TRUE)
}
\arguments{
\item{composite}{Comp object returned from CreateBrain.}

\item{legend}{Boolean whether the legend should be included. Default = TRUE}
}
\value{
plots the SGSE brain image
}
\description{
\code{PlotBrain} Plots CreateBrain. The gene set enrichment observe
within the microdissected tissues (results of testEnrich) are
combined here to show gene set enrichment across broad brain regions.
Enriched regions are colored in red, and regions depleted for the query
gene list are colored in blue.
}
\details{
PlotBrain plots your spatial gene set enrichment image.
}
\examples{

##First put together a gene list, or load in the default vth dataset
#brainImageR:::loadworkspace()
data(vth)
##Calculate the spatial enrichment.
#composite <- SpatialEnrichment(vth, reps = 20, refset = "developing")
##Calculate the significance of the gene set enrichment
#boot <- testEnrich(composite)
##Color the brain section of interest
#composite <- CreateBrain(composite, boot, slice = 5, pcut = 0.05)
##Plot the brain
#PlotBrain(composite)
}
