% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotEnrich.R
\name{PlotEnrich}
\alias{PlotEnrich}
\title{PlotEnrich}
\usage{
PlotEnrich(boot, topcut = 2, botcut = 2)
}
\arguments{
\item{boot}{results from the testEnrich function}

\item{topcut}{upper cutoff (fold-change) for plotting region names}

\item{botcut}{bottom cutoff (fold-change) for plotting region names}
}
\value{
Spatial enrichment plot
}
\description{
Plot the significance results returned from testEnrich
}
\examples{
#brainImageR:::loadworkspace()
##First put together a gene list, or load in the default vth dataset
data(vth)
##Calculate the spatial enrichment
#composite <- SpatialEnrichment(vth, 20, "developing")
#tissueExp1 <- composite@tissueExp1
#random.matrix <- composite@random.matrix
##Calculate the significance estimates
#boot <- testEnrich(composite)
#PlotEnrich(boot, topcut = 6, botcut = 0.1)

}
