% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pred.R
\docType{class}
\name{Pred-class}
\alias{Pred-class}
\alias{Pred}
\title{Set class Pred}
\description{
Set class Pred
}
\examples{
prep <- methods::new(Class="Pred",
pred_age = data.frame(matrix(0,nrow=10,ncol=10)),
model =  list(c(rep("A",5), rep("B",5))),
minage = 8,
maxage = 2120,
tissue = "HIP"
)
}
