% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PValue.onetail.R
\name{PValue.onetail}
\alias{PValue.onetail}
\title{Calculate p-value from bootstrapped sample}
\usage{
PValue.onetail(regions, tissueExp1, random.matrix)
}
\arguments{
\item{regions}{regions to search}

\item{tissueExp1}{presence of genes in query}

\item{random.matrix}{presence of genes at random}
}
\value{
p-value of the significance of tissueExp1 given the random.matrix
}
\description{
\code{PValue.onetail} Calculates the p-value from a bootstrapped sample
}
\examples{
##Internal to brainImageR, called within testEnrich
#brainImageR:::loadworkspace()
##First put together a gene list, or load in the default vth dataset
data(vth)
##Calculate the spatial enrichment.
#composite <- SpatialEnrichment(vth, reps = 20, refset = "developing")
#tissueExp1 <- composite@tissueExp1
#random.matrix <- composite@random.matrix
#boot <- PValue.onetail(regions = names(tissueExp1),
#tissueExp1,
#random.matrix)

}
