% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vth.R
\docType{data}
\name{vth}
\alias{vth}
\title{VTH genes identified from Allen Brain Atlas developing human}
\format{A vector with 1389 gene names}
\usage{
vth
}
\description{
VTH genes identified from Allen Brain Atlas developing human
}
\keyword{datasets}
